refactoring add registartion rest

+++ src/main/java/pl/tomo/service/DosageService.java
@@ -8,16 +8,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 
-import com.jcabi.aspects.Loggable;
-
+import pl.tomo.controller.exception.AccessDeniedException;
+import pl.tomo.entity.Disease;
 import pl.tomo.entity.Dosage;
+import pl.tomo.entity.Medicament;
 import pl.tomo.entity.User;
 import pl.tomo.entity.form.DosageForm;
 import pl.tomo.repository.DosageRepository;
 import pl.tomo.repository.DosageRepositoryEntityGraph;
 
 @Service
-@Loggable
 public class DosageService {
 	
 	@Autowired
@@ -43,30 +43,49 @@ public class DosageService {
 		return dosageRepository.getDosages(idMD);
 	}
 
-	public void delete(int id) {
+	public void delete(int id, HttpServletRequest request) {
+		User user = userService.findByRequestOnlyUser(request);
+		Dosage dosage = findByIdWithUser(id);
+		if(!dosage.getUser().equals(user)) 
+			throw new AccessDeniedException();
 		dosageRepository.delete(id);
 	}
 
 	public Dosage findById(int id) {
-		Dosage dosage = dosageRepositoryEntityGraph.getById("select d from Dosage d where d.id="+id, "user");
-		return dosage;
+		return dosageRepositoryEntityGraph.getById("select d from Dosage d where d.id="+id, "user");
+
+	}
+	
+	private Dosage findByIdWithUser(int id) {
+		return dosageRepositoryEntityGraph.findByIdWithUser(id);
 	}
 
 	public DosageForm getDosages(HttpServletRequest request, int idD, int idM) {
+		User user = userService.findByRequestOnlyUser(request);
+		Medicament medicament = medicamentService.findByIdWithUser(idM);
+		Disease disease = diseaseService.findByIdWithUser(idD);
+		if(!medicament.getUser().equals(user) | !disease.getUser().equals(user))
+			throw new AccessDeniedException();
 		DosageForm dosageForm = new DosageForm();
 		String sql = "select id from Disease_Medicament where disease_id=? and medicaments_id=?";
 		int idMD = jdbcTemplateMySQL.queryForObject(sql, Integer.class, idD, idM).intValue();
 		List<Dosage> dosages = getDosages(idMD);
 		dosageForm.setDosages(dosages);
-		dosageForm.setMedicament(medicamentService.findById(idM));
-		dosageForm.setDisease(diseaseService.findById(idD));
+		dosageForm.setMedicament(medicament);
+		dosageForm.setDisease(disease);
 		return dosageForm;
 	}
 
 	public Dosage save(Dosage dosage, HttpServletRequest request) {
-		String sql = "select id from Disease_Medicament where disease_id=? and medicaments_id=?";
-		int idMD = jdbcTemplateMySQL.queryForObject(sql, Integer.class, dosage.getIdD(), dosage.getIdM()).intValue();
+		int idM = dosage.getIdM();
+		Medicament medicament = medicamentService.findByIdWithUser(idM);
+		int idD = dosage.getIdD();
+		Disease disease = diseaseService.findByIdWithUser(idD);
 		User user = userService.findByRequest(request);
+		if(!medicament.getUser().equals(user) | !disease.getUser().equals(user))
+			throw new AccessDeniedException();
+		String sql = "select id from Disease_Medicament where disease_id=? and medicaments_id=?";
+		int idMD = jdbcTemplateMySQL.queryForObject(sql, Integer.class, idD, idM).intValue();
 		dosage.setIdMD(idMD);
 		dosage.setUser(user);
 		dosageRepository.save(dosage);
